Restructure libs source to support multi platform
The "libstm32l_discovery" was actually the ST std periph library, plus the
CMSIS core files.  Renamed it as such, and restructured the directories and
naming to make this easier to contain the std periph drivers for more than
just one board.

Verified with the LCD demo makefile, and loading it in sram again.